requirements.txt
# HotSwap RAG - Requirements
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.30.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Document Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# OCR (optional - requires Tesseract installation)
# pytesseract>=0.3.10
# pdf2image>=1.16.0
# easyocr>=1.7.0

# Vector Stores
chromadb>=0.4.22
# pinecone-client>=3.0.0
# pymilvus>=2.3.0
# faiss-cpu>=1.7.4

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0

# LangChain (for some vector store integrations)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Embeddings
tiktoken>=0.5.0

# Utilities
httpx>=0.26.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0